# specs/001-foundation-product-catalog/contracts/product-service-api.openapi.yaml
openapi: 3.1.0

info:
  title: API do Serviço de Produtos - TechStore
  description: Contrato REST para o Serviço de Produtos do Catálogo Fundacional TechStore
  version: 1.0.0
  contact:
    name: Suporte TechStore
    email: support@techstore.local

servers:
  - url: http://localhost:8080/api
    description: Ambiente de Desenvolvimento
  - url: https://api.techstore.com
    description: Ambiente de Produção

paths:
  /products:
    get:
      summary: Listar todos os produtos
      description: Retorna uma lista paginada de produtos com filtros e opciones de ordenação opcionais
      operationId: listProducts
      tags:
        - Produtos
      parameters:
        - name: query
          in: query
          description: Termo de busca full-text (pesquisa em nome, marca, descrição)
          schema:
            type: string
        - name: categoryId
          in: query
          description: UUID da categoria para filtrar produtos
          schema:
            type: string
            format: uuid
        - name: minPrice
          in: query
          description: Preço mínimo (inclusive)
          schema:
            type: number
            format: double
            minimum: 0
        - name: maxPrice
          in: query
          description: Preço máximo (inclusive)
          schema:
            type: number
            format: double
            minimum: 0
        - name: inStock
          in: query
          description: Filtrar apenas produtos em estoque (quantidade > 0)
          schema:
            type: boolean
        - name: brand
          in: query
          description: Filtrar por marca específica
          schema:
            type: string
        - name: sortBy
          in: query
          description: Campo para ordenação
          schema:
            type: string
            enum: [relevance, price, name, newest]
            default: relevance
        - name: sortOrder
          in: query
          description: Direção da ordenação
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          description: Número da página (0-based)
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: pageSize
          in: query
          description: Número de itens por página
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista paginada de produtos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductsResponse'
              example:
                content:
                  - id: 550e8400-e29b-41d4-a716-446655440000
                    name: MacBook Pro 16
                    description: Laptop de alta performance
                    price: 2499.99
                    brand: Apple
                    sku: SKU-001
                    categoryId: 550e8400-e29b-41d4-a716-446655440001
                    quantity: 50
                    imageUrl: https://example.com/macbook.jpg
                    isActive: true
                    createdAt: "2026-01-15T10:30:00Z"
                    updatedAt: "2026-01-20T14:45:00Z"
                totalElements: 1234
                totalPages: 62
                currentPage: 0
                pageSize: 20
                hasMore: true
        '400':
          description: Parâmetros de requisição inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Criar novo produto
      description: Cria um novo produto no catálogo. Requer autorização ADMIN.
      operationId: createProduct
      tags:
        - Produtos
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            example:
              name: Dell XPS 15
              description: Laptop premium ultrabook
              price: 1999.99
              cost: 1200.00
              brand: Dell
              sku: SKU-DELLXPS15
              categoryId: 550e8400-e29b-41d4-a716-446655440002
              quantity: 25
              imageUrl: https://example.com/dell-xps-15.jpg
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Dados de requisição inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada (requer role ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflito (SKU ou nome+categoria já existem)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/{id}:
    get:
      summary: Obter detalhes de um produto
      description: Retorna os detalhes completos de um produto específico
      operationId: getProductById
      tags:
        - Produtos
      parameters:
        - name: id
          in: path
          required: true
          description: UUID do produto
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhes do produto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Atualizar um produto
      description: Atualiza os detalhes de um produto existente. Requer autorização ADMIN.
      operationId: updateProduct
      tags:
        - Produtos
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID do produto
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
            example:
              name: MacBook Pro 16 (2026)
              price: 2799.99
              quantity: 40
      responses:
        '200':
          description: Produto atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Dados de requisição inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada (requer role ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflito (SKU ou nome+categoria já existem)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Deletar um produto (soft delete)
      description: Marca um produto como inativo (soft delete). Requer autorização ADMIN.
      operationId: deleteProduct
      tags:
        - Produtos
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID do produto
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Produto deletado com sucesso (sem conteúdo)
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada (requer role ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /categories:
    get:
      summary: Listar todas as categorias
      description: Retorna a lista completa de categorias
      operationId: listCategories
      tags:
        - Categorias
      responses:
        '200':
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDTO'
              example:
                - id: 550e8400-e29b-41d4-a716-446655440001
                  name: Eletrônicos
                  description: Dispositivos eletrônicos e gadgets
                  parentCategoryId: null
                  displayOrder: 0
                  isActive: true
                  createdAt: "2026-01-15T10:30:00Z"
                  updatedAt: "2026-01-15T10:30:00Z"
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Criar nova categoria
      description: Cria uma nova categoria. Requer autorização ADMIN.
      operationId: createCategory
      tags:
        - Categorias
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
            example:
              name: Computadores
              description: Laptops, desktops, tablets
              parentCategoryId: null
      responses:
        '201':
          description: Categoria criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDTO'
        '400':
          description: Dados de requisição inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada (requer role ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflito (nome da categoria já existe)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /categories/{id}:
    get:
      summary: Obter detalhes de uma categoria
      description: Retorna os detalhes de uma categoria específica
      operationId: getCategoryById
      tags:
        - Categorias
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da categoria
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhes da categoria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDTO'
        '404':
          description: Categoria não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Atualizar uma categoria
      description: Atualiza os detalhes de uma categoria existente. Requer autorização ADMIN.
      operationId: updateCategory
      tags:
        - Categorias
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da categoria
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
            example:
              name: Eletrônicos Pessoais
              description: Dispositivos eletrônicos pessoais e gadgets
      responses:
        '200':
          description: Categoria atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDTO'
        '400':
          description: Dados de requisição inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada (requer role ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Categoria não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflito (nome da categoria já existe)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Deletar uma categoria (soft delete)
      description: Marca uma categoria como inativa (soft delete). Requer autorização ADMIN.
      operationId: deleteCategory
      tags:
        - Categorias
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da categoria
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Categoria deletada com sucesso (sem conteúdo)
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada (requer role ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Categoria não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via endpoint de autenticação. Use no header Authorization como "Bearer {token}"

  schemas:
    ProductDTO:
      type: object
      description: Representação de um produto retornado pela API
      required:
        - id
        - name
        - price
        - sku
        - categoryId
        - quantity
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único do produto
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Nome do produto
        description:
          type: string
          maxLength: 5000
          description: Descrição detalhada do produto
          nullable: true
        price:
          type: number
          format: double
          minimum: 0
          description: Preço atual em moeda padrão
        cost:
          type: number
          format: double
          minimum: 0
          description: Custo interno do produto
          nullable: true
        brand:
          type: string
          maxLength: 100
          description: Nome da marca do produto
          nullable: true
        sku:
          type: string
          maxLength: 50
          description: Stock Keeping Unit (único no sistema)
        categoryId:
          type: string
          format: uuid
          description: UUID da categoria a que o produto pertence
        quantity:
          type: integer
          minimum: 0
          description: Quantidade em estoque
        imageUrl:
          type: string
          format: uri
          maxLength: 2000
          description: URL da imagem primária do produto
          nullable: true
        isActive:
          type: boolean
          description: Indica se o produto está ativo
        createdAt:
          type: string
          format: date-time
          description: Timestamp de criação (UTC, somente leitura)
        updatedAt:
          type: string
          format: date-time
          description: Timestamp de última atualização (UTC, somente leitura)

    CreateProductRequest:
      type: object
      description: Dados obrigatórios para criar um novo produto
      required:
        - name
        - price
        - sku
        - categoryId
        - quantity
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Nome do produto
        description:
          type: string
          maxLength: 5000
          description: Descrição detalhada do produto
        price:
          type: number
          format: double
          minimum: 0
          description: Preço em moeda padrão
        cost:
          type: number
          format: double
          minimum: 0
          description: Custo interno
        brand:
          type: string
          maxLength: 100
          description: Nome da marca
        sku:
          type: string
          maxLength: 50
          description: Stock Keeping Unit (deve ser único)
        categoryId:
          type: string
          format: uuid
          description: UUID da categoria
        quantity:
          type: integer
          minimum: 0
          description: Quantidade inicial em estoque
        imageUrl:
          type: string
          format: uri
          maxLength: 2000
          description: URL da imagem primária

    UpdateProductRequest:
      type: object
      description: Dados para atualizar um produto existente
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Nome do produto
        description:
          type: string
          maxLength: 5000
          description: Descrição detalhada
        price:
          type: number
          format: double
          minimum: 0
          description: Preço em moeda padrão
        cost:
          type: number
          format: double
          minimum: 0
          description: Custo interno
        brand:
          type: string
          maxLength: 100
          description: Nome da marca
        quantity:
          type: integer
          minimum: 0
          description: Quantidade em estoque
        imageUrl:
          type: string
          format: uri
          maxLength: 2000
          description: URL da imagem

    CategoryDTO:
      type: object
      description: Representação de uma categoria
      required:
        - id
        - name
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único da categoria
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Nome da categoria (único)
        description:
          type: string
          maxLength: 500
          description: Descrição da categoria
          nullable: true
        parentCategoryId:
          type: string
          format: uuid
          description: UUID da categoria pai (para hierarquia)
          nullable: true
        displayOrder:
          type: integer
          description: Ordem de exibição
        isActive:
          type: boolean
          description: Indica se a categoria está ativa
        createdAt:
          type: string
          format: date-time
          description: Timestamp de criação (UTC, somente leitura)
        updatedAt:
          type: string
          format: date-time
          description: Timestamp de última atualização (UTC, somente leitura)

    CreateCategoryRequest:
      type: object
      description: Dados obrigatórios para criar uma nova categoria
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Nome da categoria (deve ser único)
        description:
          type: string
          maxLength: 500
          description: Descrição da categoria
        parentCategoryId:
          type: string
          format: uuid
          description: UUID da categoria pai (opcional)

    UpdateCategoryRequest:
      type: object
      description: Dados para atualizar uma categoria existente
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Nome da categoria
        description:
          type: string
          maxLength: 500
          description: Descrição da categoria
        displayOrder:
          type: integer
          description: Ordem de exibição

    PaginatedProductsResponse:
      type: object
      description: Resposta paginada de lista de produtos
      required:
        - content
        - totalElements
        - totalPages
        - currentPage
        - pageSize
        - hasMore
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductDTO'
          description: Lista de produtos na página atual
        totalElements:
          type: integer
          description: Número total de produtos (todos)
        totalPages:
          type: integer
          description: Número total de páginas
        currentPage:
          type: integer
          description: Número da página atual (0-based)
        pageSize:
          type: integer
          description: Número de itens por página
        hasMore:
          type: boolean
          description: Indica se há mais páginas

    ErrorResponse:
      type: object
      description: Resposta de erro padronizada
      required:
        - status
        - message
        - correlationId
      properties:
        status:
          type: integer
          description: Código HTTP de status
        message:
          type: string
          description: Mensagem de erro legível
        details:
          type: string
          description: Detalhes adicionais do erro
          nullable: true
        correlationId:
          type: string
          description: ID único para rastreamento entre serviços
        timestamp:
          type: string
          format: date-time
          description: Timestamp do erro (UTC)
          nullable: true